The front end is still a bit of a mess, I prioritized working on the backend and the connection between the server, client and database, which I feel like I have made a lot of progess on.  The front end is easier to troubleshoot and not conceptually difficult, but time consuming for me to sift through the docs to find the proper way of actualizing my intended goal. This is something I am sure I can figure out on my own, I just need more practice whereas the backend is something that, before this, I felt I needed assistance with.
Database tests work
API is set up but I need to refactor to now_playing
Express server calls database and is set up for API

I realize that I was supposed to finish this, I really did put in a lot of hours, and whatever the case I still do intend to finish it.  I think I will be stronger on thesis, because now that I've figured out how to navigate through the server and database a bit more easily, the logic inside of it is more intuitive, than the front end.  Let me know if there is anything else I can do.  Thank you.